+++ tester/tests/modules/router/fork/fork-context-tester.cc
@@ -100,8 +100,7 @@ void notRtpPortAndForkCallContext() {
 	Server proxy{kConfig};
 	proxy.setConfigParameter({"module::Router/fork-late", "false"});
 	proxy.setConfigParameter({"module::MediaRelay/enabled", "true"});
-	proxy.setConfigParameter({"module::MediaRelay/sdp-port-range-min", "1023"});
-	proxy.setConfigParameter({"module::MediaRelay/sdp-port-range-max", "1024"});
+	proxy.setConfigParameter({"module::MediaRelay/sdp-port-range", "1023-1024"});
 	proxy.start();
 
 	const auto& registrarDb = proxy.getRegistrarDb();
@@ -425,8 +424,7 @@ public:
 	nta_agent_t* getSofiaAgent() const override {
 		return nullptr;
 	}
-	void injectRequest(unique_ptr<RequestSipEvent>&&) override {
-	}
+	void injectRequest(unique_ptr<RequestSipEvent>&&) override {}
 	unique_ptr<ResponseSipEvent> injectResponse(unique_ptr<ResponseSipEvent>&& ev) override {
 		return std::move(ev);
 	}
@@ -481,8 +479,7 @@ class FindBestBranchTest : public Test {
 public:
 	FindBestBranchTest(bool brFound, int statusCodeExpected, const vector<int>& statusList, bool avoid503And408 = true)
 	    : mBrFound(brFound), mStatusCodeExpected(statusCodeExpected), mStatusList(statusList),
-	      mAvoid503And408(avoid503And408) {
-	}
+	      mAvoid503And408(avoid503And408) {}
 	void operator()() override {
 		ForkContextForTest fork{mAgentMock.get()};
 		for_each(mStatusList.begin(), mStatusList.end(), [&fork](auto i) {